Overnight maintenance: cross-flavor parity fixes + spec/ADR accuracy (#55)
62 commits across all four flavors + spec/ADRs/examples/docs. Behavioral fixes (async external-cancel re-raise, C# TokenPagedComposition CanExecuteChanged trigger, Swift scorer Int->Double / removeAt sync-clear / SearchableState post-dispose / RelayCommandOf.dispose, empty-clear Reset gating, TS forwarding setter, new Swift collection+hub API), spec/ADR accuracy corrections (stale Swift-subset prose, spec/11 dispatcher table), and ADR-0009 divergence documentation. All flavors + examples green; codebase converged.

+++ langs/swift/Sources/VMx/Composites/ScoredFilteredCompositeVM.swift
@@ -3,12 +3,12 @@
 //
 
 public final class ScoredFilteredCompositeVM<Child: ComponentVMBase>: FilteredCompositeVM<Child> {
-    private let scorer: (Child) -> Int?
+    private let scorer: (Child) -> Double?
 
     public init(
         _ source: CompositeVM<Child>,
         cursorPolicy: FilteredCursorPolicy = .snapToFirst,
-        scorer: @escaping (Child) -> Int?
+        scorer: @escaping (Child) -> Double?
     ) {
         self.scorer = scorer
         super.init(
